+++ packages/atomic/src/components/insight/atomic-insight-interface/atomic-insight-interface.tsx
@@ -10,12 +10,12 @@ import {
   Watch,
 } from '@stencil/core';
 import i18next, {i18n} from 'i18next';
-import {InitializeEvent} from '../../../utils/initialization-utils';
-import {CommonBindings} from '../../common/interface/bindings';
+import {InitializeEvent} from '@utils/initialization-utils';
+import {CommonBindings} from '@components/common/interface/bindings';
 import {
   BaseAtomicInterface,
   CommonAtomicInterfaceHelper,
-} from '../../common/interface/interface-common';
+} from '@components/common/interface/interface-common';
 import {createAtomicSvcInsightStore} from './store';
 
 const FirstInsightRequestExecutedFlag = 'firstInsightRequestExecuted';