+++ packages/atomic/src/components/insight/atomic-insight-interface/store.ts
@@ -2,7 +2,7 @@ import {InsightEngine} from '@coveo/headless/insight';
 import {
   AtomicCommonStoreData,
   createAtomicCommonStore,
-} from '../../common/interface/store';
+} from '@components/common/interface/store';
 
 export interface AtomicSvgInsightStoreData extends AtomicCommonStoreData {}