+++ packages/atomic-react/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.13.4](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@1.13.3...@coveo/atomic-react@1.13.4) (2022-06-30)
+
+**Note:** Version bump only for package @coveo/atomic-react
+
+
+
+
+
+## [1.13.3](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@1.13.2...@coveo/atomic-react@1.13.3) (2022-06-30)
+
+**Note:** Version bump only for package @coveo/atomic-react
+
+
+
+
+
 ## [1.13.2](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@1.13.1...@coveo/atomic-react@1.13.2) (2022-06-29)
 
 **Note:** Version bump only for package @coveo/atomic-react