+++ packages/bueno/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.40.4](https://github.com/coveo/ui-kit/compare/@coveo/bueno@0.40.3...@coveo/bueno@0.40.4) (2022-06-30)
+
+**Note:** Version bump only for package @coveo/bueno
+
+
+
+
+
+## [0.40.3](https://github.com/coveo/ui-kit/compare/@coveo/bueno@0.40.2...@coveo/bueno@0.40.3) (2022-06-30)
+
+**Note:** Version bump only for package @coveo/bueno
+
+
+
+
+
 ## [0.40.2](https://github.com/coveo/ui-kit/compare/@coveo/bueno@0.40.1...@coveo/bueno@0.40.2) (2022-06-29)
 
 **Note:** Version bump only for package @coveo/bueno